+++ build-dev.sh
@@ -26,6 +26,7 @@ cp -r static build/
 cp -r GeoLite2-Country.mmdb build/
 cp nebula-arm-cert nebula-cert nebula-arm nebula build/
 cp -r Logo.png build/
+cp restic build/
 mkdir build/images
 cp client/src/assets/images/icons/cosmos_gray.png build/cosmos_gray.png
 cp client/src/assets/images/icons/cosmos_gray.png cosmos_gray.png