+++ build.sh
@@ -36,6 +36,7 @@ chmod +x build/cosmos-launcher-arm64
 cp -r static build/
 cp -r GeoLite2-Country.mmdb build/
 cp nebula-arm-cert nebula-cert nebula-arm nebula build/
+cp restic-arm restic build/
 cp -r Logo.png build/
 mkdir build/images
 cp client/src/assets/images/icons/cosmos_gray.png build/cosmos_gray.png