+++ client/src/api/backup.ts
@@ -1,0 +1,76 @@
+import wrap from './wrap';
+
+interface BackupConfig {
+  name: string;
+  source: string;
+  repository: string;
+  password: string;
+  crontab?: string;
+  tags?: string[];
+  exclude?: string[];
+}
+
+interface RestoreConfig {
+  snapshotId: string;
+  target: string;
+  include?: string[];
+}
+
+function listSnapshots(name: string) {
+  return wrap(fetch(`/cosmos/api/backups/${name}/snapshots`, {
+    method: 'GET',
+    headers: {
+      'Content-Type': 'application/json'
+    },
+  }))
+}
+
+function listFolders(name: string, snapshot: string, path?: string) {
+  return wrap(fetch(`/cosmos/api/backups/${name}/folders?path=${path || ''}`, {
+    method: 'GET',
+    headers: {
+      'Content-Type': 'application/json'
+    },
+    body: JSON.stringify({ snapshot })
+  }))
+}
+
+function restoreBackup(name: string, config: RestoreConfig) {
+  return wrap(fetch(`/cosmos/api/backups/${name}/restore`, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json'
+    },
+    body: JSON.stringify(config)
+  }))
+}
+
+function addBackup(config: BackupConfig) {
+  return wrap(fetch('/cosmos/api/backups', {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json'
+    },
+    body: JSON.stringify(config)
+  }))
+}
+
+function removeBackup(name: string, deleteRepo: boolean = false) {
+  return wrap(fetch(`/cosmos/api/backups/${name}`, {
+    method: 'DELETE',
+    headers: {
+      'Content-Type': 'application/json'
+    },
+    body: JSON.stringify({ deleteRepo })
+  }))
+}
+
+export {
+  listSnapshots,
+  listFolders,
+  restoreBackup,
+  addBackup,
+  removeBackup,
+  BackupConfig,
+  RestoreConfig
+};