+++ src/proxy/shield.go
@@ -314,7 +314,7 @@ func GetClientID(r *http.Request, route utils.ProxyRouteConfig) string {
 }
 
 func isPrivileged(req *http.Request, policy utils.SmartShieldPolicy) bool {
-	role, _ := strconv.Atoi(req.Header.Get("x-cosmos-role"))
+	role, _ := strconv.Atoi(req.Header.Get("x-cosmos-user-role"))
 	return role >= policy.PrivilegedGroups
 }