+++ changelog.md
@@ -1,4 +1,5 @@
 ## Version 0.18.0
+ - UI to backup and restore containers/folders/volumes using Restic
  - Implements sudo mode - your normal token last longer, but you need to "sudo" to do admin tasks
  - Re-Implements the SSO using openID internally - fixes issue where you need to re-loging when app are on different domains (because of browser cookies limitations)
  - Implements local HTTPS Certificate Authority, to locally trust self-signed certificates on devices
@@ -7,6 +8,7 @@
  - Localizations improvements (Thanks @madejackson)
  - Improved local IP detection (Thanks @r41d)
  - Updated LEGO to 4.21.0
+ - Fix RClone not starting
 
 ## Version 0.17.7
  - Fix error code on login screen